get_name on Exceptions should return the type name.
Previously it raised TypeError because you can't create a weak reference
to an Exception type. This change allows the use of `get_name` on
Exceptions, returning the type name instead of raising an error.

Other types which are not weak referenceable will will return a name
constructed from the type name and the id() of the object.